Prep for mainnet deploy (#458)
