+++ src/test/mod.rs
@@ -1,5 +1,7 @@
 // Test related stuff
 
+pub mod node_e2e;
+
 fn asdf() {
 
 }
